+++ jwp-basic-gradle/src/main/java/core/di/factory/AnnotatedBeanDefinitionReader.java
@@ -1,4 +1,4 @@
-package core.di;
+package core.di.factory;
 
 import core.annotation.Bean;
 import java.lang.reflect.Method;
