+++ jwp-basic-gradle/src/main/java/core/di/factory/ApplicationContext.java
@@ -1,4 +1,4 @@
-package core.di;
+package core.di.factory;
 
 import java.util.Set;
