+++ jwp-basic-gradle/src/main/java/core/di/factory/BeanDefinition.java
@@ -1,4 +1,4 @@
-package core.di;
+package core.di.factory;
 
 import com.google.common.collect.Sets;
 import java.lang.reflect.Constructor;
