+++ jwp-basic-gradle/src/main/java/core/di/factory/BeanDefinitionRegistry.java
@@ -1,4 +1,4 @@
-package core.di;
+package core.di.factory;
 
 public interface BeanDefinitionRegistry {
