+++ jwp-basic-gradle/src/main/java/core/di/factory/ClasspathBeanDefinitionScanner.java
@@ -1,4 +1,4 @@
-package core.di;
+package core.di.factory;
 
 import com.google.common.collect.Sets;
 import core.annotation.Component;
