+++ jwp-basic-gradle/src/main/java/core/di/factory/InjectType.java
@@ -1,4 +1,4 @@
-package core.di;
+package core.di.factory;
 
 public enum InjectType {
     INJECT_NO, INJECT_CONSTRUCTOR, INJECT_FIELD
