+++ jwp-basic-gradle/src/test/java/core/nmvc/AnnotationHandlerMappingTest.java
@@ -2,7 +2,7 @@ package core.nmvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import core.di.AnnotationConfigApplicationContext;
+import core.di.factory.AnnotationConfigApplicationContext;
 import core.web.mvcframework.ModelAndView;
 import core.web.mvcframework.mapping.annotation.AnnotationHandlerMapping;
 import core.web.mvcframework.mapping.annotation.HandlerExecution;
